backend/requirements.txt
Flask>=2.1
Flask-Login>=0.6
Flask-SQLAlchemy>=3.0
PyMySQL>=1.0
python-dotenv>=0.21
passlib[bcrypt]>=1.7
Flask-Cors>=3.0.10
marshmallow>=3.0
PyJWT>=2.0
cryptography>=40.0.0
bcrypt>=3.2.2
reportlab>=4.0
openpyxl>=3.1
